Refactor: Rename package and classes
Remove unused parameter and method
