+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/user/accounts/create/data/PaymentDataModel.java
@@ -15,19 +15,28 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.main.content.user.accounts.create.account_data.method_forms;
+package bisq.desktop.main.content.user.accounts.create.data;
 
+import bisq.account.payment_method.PaymentMethod;
 import bisq.desktop.common.view.Model;
-import javafx.beans.property.BooleanProperty;
-import javafx.beans.property.SimpleBooleanProperty;
+import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleObjectProperty;
 import lombok.Getter;
 
 import java.util.HashMap;
 import java.util.Map;
 
-@Getter
-public class PaymentFormModel implements Model {
-    private final Map<String, Object> formData = new HashMap<>();
-    private final Map<String, String> validationErrors = new HashMap<>();
-    private final BooleanProperty isValid = new SimpleBooleanProperty(false);
+public class PaymentDataModel implements Model {
+    @Getter
+    private final ObjectProperty<PaymentMethod<?>> paymentMethod = new SimpleObjectProperty<>();
+    @Getter
+    private final ObjectProperty<Map<String, Object>> accountData = new SimpleObjectProperty<>(new HashMap<>());
+
+    public ObjectProperty<PaymentMethod<?>> paymentMethodProperty() {
+        return paymentMethod;
+    }
+
+    public void setPaymentMethod(PaymentMethod<?> paymentMethod) {
+        this.paymentMethod.set(paymentMethod);
+    }
 }
